+++ src/PVOutput.Net/Objects/IFavourite.cs
@@ -63,7 +63,7 @@ namespace PVOutput.Net.Objects
         /// <summary>
         /// Primary orientation of most panels.
         /// </summary>
-        string Orientation { get; set; }
+        Orientation Orientation { get; set; }
 
         /// <summary>
         /// Tilt the solar array has.