+++ src/PVOutput.Net/Objects/Modules/Readers/SystemObjectStringReader.cs
@@ -34,7 +34,7 @@ namespace PVOutput.Net.Objects.Modules.Readers
                 (t, s) => t.NumberOfInverters = FormatHelper.GetValueOrDefault<int>(s),
                 (t, s) => t.InverterPower = FormatHelper.GetValueOrDefault<int>(s),
                 (t, s) => t.InverterBrand = s,
-                (t, s) => t.Orientation = s,
+                (t, s) => t.Orientation = FormatHelper.DescriptionToEnumValue<Orientation>(s),
                 (t, s) => t.ArrayTilt = FormatHelper.GetValue<decimal>(s),
                 (t, s) => t.Shade = !string.IsNullOrWhiteSpace(s) ? FormatHelper.DescriptionToEnumValue<Shade>(s) : Shade.None,
                 (t, s) => t.InstallDate = FormatHelper.ParseOptionalDate(s),
@@ -43,7 +43,7 @@ namespace PVOutput.Net.Objects.Modules.Readers
                 (t, s) => t.StatusInterval = FormatHelper.GetValueOrDefault<int>(s),
                 (t, s) => t.SecondaryNumberOfPanels = FormatHelper.GetValue<int>(s),
                 (t, s) => t.SecondaryPanelPower = FormatHelper.GetValue<int>(s),
-                (t, s) => t.SecondaryOrientation = s,
+                (t, s) => t.SecondaryOrientation = !string.IsNullOrWhiteSpace(s) ? FormatHelper.DescriptionToEnumValue<Orientation>(s) : (Orientation?)null,
                 (t, s) => t.SecondaryArrayTilt = FormatHelper.GetValue<decimal>(s)
             };