+++ tests/PVOutput.Net.Tests/Modules/Search/SearchServiceTests.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using NUnit.Framework;
+using PVOutput.Net.Enums;
 using PVOutput.Net.Modules;
 using PVOutput.Net.Objects;
 using PVOutput.Net.Objects.Modules.Readers;
@@ -201,7 +202,7 @@ namespace PVOutput.Net.Tests.Modules.Search
                 Assert.That(result.SystemSize, Is.EqualTo(9360));
                 Assert.That(result.Postcode, Is.EqualTo(4280));
                 Assert.That(result.Country, Is.EqualTo("Australia"));
-                Assert.That(result.Orientation, Is.EqualTo("NW"));
+                Assert.That(result.Orientation, Is.EqualTo(Orientation.NorthWest));
                 Assert.That(result.NumberOfOutputs, Is.EqualTo(81));
                 Assert.That(result.LastOutput, Is.EqualTo("2 days ago"));
                 Assert.That(result.SystemId, Is.EqualTo(249));